Switch to downloading test-butler APK in jest's global-setup
