--- detox/test/detox.config-raw.json
@@ -1,86 +1,0 @@
-{
-  "test-runner": "nyc jest",
-  "runner-config": "e2e/config.js",
-  "behavior": {
-    "init": {
-      "reinstallApp": true,
-      "launchApp": true,
-      "exposeGlobals": true
-    },
-    "cleanup": {
-      "shutdownDevice": false
-    }
-  },
-  "__session": {
-    "server": "ws://localhost:8099",
-    "sessionId": "test"
-  },
-  "artifacts": {
-    "plugins": {
-      "log": "all",
-      "screenshot": {
-        "shouldTakeAutomaticSnapshots": true,
-        "takeWhen": {
-          "testDone": true
-        }
-      },
-      "timeline": "all"
-    }
-  },
-  "configurations": {
-    "ios.sim.debug": {
-      "binaryPath": "ios/build/Build/Products/Debug-iphonesimulator/example.app",
-      "build": "set -o pipefail && xcodebuild -workspace ios/example.xcworkspace -UseNewBuildSystem=NO -scheme example_ci -configuration Debug -sdk iphonesimulator -derivedDataPath ios/build -quiet",
-      "type": "ios.simulator",
-      "device": {
-        "type": "iPhone 11 Pro"
-      }
-    },
-    "ios.sim.release": {
-      "binaryPath": "ios/build/Build/Products/Release-iphonesimulator/example.app",
-      "build": "set -o pipefail && export CODE_SIGNING_REQUIRED=NO && export RCT_NO_LAUNCH_PACKAGER=true && xcodebuild -workspace ios/example.xcworkspace -UseNewBuildSystem=NO -scheme example_ci -configuration Release -sdk iphonesimulator -derivedDataPath ios/build -quiet",
-      "type": "ios.simulator",
-      "device": {
-        "type": "iPhone 11 Pro"
-      }
-    },
-    "ios.none": {
-      "binaryPath": "ios",
-      "type": "ios.none",
-      "device": {
-        "type": "iPhone 11 Pro"
-      },
-      "session": {
-        "server": "ws://localhost:8099",
-        "sessionId": "com.wix.detox-example"
-      }
-    },
-    "android.emu.debug": {
-      "binaryPath": "android/app/build/outputs/apk/fromBin/debug/app-fromBin-debug.apk",
-      "utilBinaryPaths": ["./cache/test-butler-app.apk"],
-      "build": "cd android && ./gradlew assembleFromBinDebug assembleFromBinDebugAndroidTest -DtestBuildType=debug && cd ..",
-      "type": "android.emulator",
-      "device": {
-        "avdName": "Pixel_API_28"
-      }
-    },
-    "android.emu.release": {
-      "binaryPath": "android/app/build/outputs/apk/fromBin/release/app-fromBin-release.apk",
-      "utilBinaryPaths": ["./cache/test-butler-app.apk"],
-      "build": "cd android && ./gradlew assembleFromBinRelease assembleFromBinReleaseAndroidTest -DtestBuildType=release && cd ..",
-      "type": "android.emulator",
-      "device": {
-        "avdName": "Pixel_API_28"
-      }
-    },
-    "android.emu.debug.fromSource": {
-      "binaryPath": "android/app/build/outputs/apk/fromSource/debug/app-fromSource-debug.apk",
-      "utilBinaryPaths": ["./cache/test-butler-app.apk"],
-      "build": "cd android && ./gradlew assembleFromSourceDebug assembleFromSourceDebugAndroidTest -DtestBuildType=debug && cd ..",
-      "type": "android.emulator",
-      "device": {
-        "avdName": "Pixel_API_28"
-      }
-    }
-  }
-}
