+++ detox/test/e2e/config.js
@@ -3,6 +3,7 @@ module.exports = {
   "testEnvironment": "./test/e2e/environment.js",
   "testRunner": "./test/node_modules/jest-circus/runner",
   "setupFilesAfterEnv": ['./test/e2e/init-coverage.js'],
+  "globalSetup": "./test/e2e/global-setup.js",
   "testTimeout": 120000,
   "reporters": process.env.DISABLE_JUNIT_REPORTER === '1'
     ? ["<rootDir>/runners/jest/streamlineReporter"]
